Moved the unicorn wrapper outside of the server.py
Lots of code tidying up to get ready for merging to master.  Moved the
unicorn wrapper class to its own file, added some additional utility
methods for handling HSV (integer) and HTML colour profiles and created
a method to set the unicorn to a specific colour.  I hope this will make
the class more useful for people and make it more reuseable.  Would like
to create a dummy unicorn which displays on screen the output either in
the console or as a TK window.  Things to think about but in anyevent
will become a new branch before merging to master.

Switched from /api/switch to /api/display.  Added a displayRainbow method.
This should provide a similar effect to Razer Synapses spectrum effect.
Considering adding other effects as well but I think thats again things
for a new branch