Massive update to README and install.sh
This is a housekeeping commit.  Some of the files have been moved around
a bit to make more sense.  Some bug fixes have been done to the
server.py due to testing during the writing of the README.md.  The
README.md is now pretty complete with a full API reference.  Finally
the install script has been completley rewritten to allow for repeatable
installation and installation via a curl request.